+++ lib/feature_flag_kit.dart
@@ -10,3 +10,6 @@
 library;
 
 export 'src/bucketing.dart';
+export 'src/models/evaluation_result.dart';
+export 'src/models/remote_config.dart';
+export 'src/models/user_context.dart';